Backend/requirements.txt
fastapi==0.101.1
PyJWT==2.8.0
motor==3.2.0
passlib==1.7.4
Pillow==10.0.0
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
uvicorn==0.23.2

